+++ app/controllers/api/v2/account_lists/donations_controller.rb
@@ -85,7 +85,7 @@ class Api::V2::AccountLists::DonationsController < Api::V2Controller
   end
 
   def permitted_filters
-    []
+    [:contact_id]
   end
 
   def pundit_user