+++ app/serializers/organization_serializer.rb
@@ -1,0 +1,8 @@
+class OrganizationSerializer < ApplicationSerializer
+  attributes :abbreviation,
+             :code,
+             :country,
+             :default_currency_code,
+             :logo,
+             :name
+end